refactor(basectl): move read-only commands into library
