+++ bin/basectl/Cargo.toml
@@ -15,10 +15,7 @@ workspace = true
 url = { workspace = true }
 anyhow = { workspace = true }
 tracing = { workspace = true }
-alloy-eips = { workspace = true }
 basectl-cli = { workspace = true }
-alloy-provider = { workspace = true }
-base-common-network = { workspace = true }
 base-consensus-peers = { workspace = true }
 base-prover-service-protocol = { workspace = true }
 rustls = { workspace = true, features = ["ring"] }
@@ -26,8 +23,6 @@ clap = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true, features = ["std"] }
 serde = { workspace = true, features = ["derive", "std"] }
 alloy-primitives = { workspace = true, features = ["serde"] }
-alloy-rpc-types-eth = { workspace = true, features = ["std"] }
-base-protocol = { workspace = true, features = ["serde", "std"] }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "time"] }
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt"] }
